refactor: clean up FiatValue props
