+++ src/common/pure/CurrencyAmountPreview/index.tsx
@@ -1,6 +1,6 @@
 import * as styledEl from '../CurrencyInputPanel/styled'
 import { CurrencySelectButton } from 'common/pure/CurrencySelectButton'
-import { FiatValue } from 'legacy/components/CurrencyInputPanel/FiatValue/FiatValueMod'
+import { FiatValue } from 'legacy/components/CurrencyInputPanel/FiatValue'
 import { Trans } from '@lingui/macro'
 import { PriceImpact } from 'legacy/hooks/usePriceImpact'
 import { CurrencyInfo } from 'common/pure/CurrencyInputPanel/types'
@@ -19,7 +19,6 @@ export interface CurrencyPreviewProps extends Partial<BuiltItProps> {
 
 export function CurrencyAmountPreview(props: CurrencyPreviewProps) {
   const { id, currencyInfo, className, priceImpactParams, topLabel } = props
-  const { priceImpact, loading: priceImpactLoading } = priceImpactParams || {}
   const { currency, balance, fiatAmount, amount } = currencyInfo
 
   return (
@@ -48,7 +47,7 @@ export function CurrencyAmountPreview(props: CurrencyPreviewProps) {
           </div>
           <div>
             <styledEl.FiatAmountText>
-              <FiatValue priceImpactLoading={priceImpactLoading} fiatValue={fiatAmount} priceImpact={priceImpact} />
+              <FiatValue fiatValue={fiatAmount} priceImpactParams={priceImpactParams} />
             </styledEl.FiatAmountText>
           </div>
         </styledEl.CurrencyInputBox>
