+++ src/common/pure/CurrencyInputPanel/CurrencyInputPanel.tsx
@@ -3,7 +3,7 @@ import * as styledEl from './styled'
 import { CurrencySelectButton } from 'common/pure/CurrencySelectButton'
 import { Currency } from '@uniswap/sdk-core'
 import CurrencySearchModal from 'legacy/components/SearchModal/CurrencySearchModal'
-import { FiatValue } from 'legacy/components/CurrencyInputPanel/FiatValue/FiatValueMod'
+import { FiatValue } from 'legacy/components/CurrencyInputPanel/FiatValue'
 import { Trans } from '@lingui/macro'
 import { PriceImpact } from 'legacy/hooks/usePriceImpact'
 import { ReceiveAmount } from 'modules/swap/pure/ReceiveAmount'
@@ -66,7 +66,6 @@ export function CurrencyInputPanel(props: CurrencyInputPanelProps) {
   } = props
 
   const isSupportedNetwork = isSupportedChainId(props.chainId as number | undefined)
-  const { priceImpact, loading: priceImpactLoading } = priceImpactParams || {}
   const { field, currency, balance, fiatAmount, amount, isIndependent, receiveAmountInfo } = currencyInfo
   const disabled = props.disabled || !isSupportedNetwork
   const viewAmount = formatInputAmount(amount, balance, isIndependent)
@@ -154,12 +153,7 @@ export function CurrencyInputPanel(props: CurrencyInputPanelProps) {
           </div>
           <div>
             <styledEl.FiatAmountText>
-              <FiatValue
-                isLoading={isRateLoading}
-                priceImpactLoading={priceImpactLoading}
-                fiatValue={fiatAmount}
-                priceImpact={priceImpact}
-              />
+              <FiatValue priceImpactParams={priceImpactParams} fiatValue={isRateLoading ? null : fiatAmount} />
             </styledEl.FiatAmountText>
           </div>
         </styledEl.CurrencyInputBox>
