+++ src/legacy/components/swap/SwapModalHeader/SwapModalHeaderMod.tsx
@@ -163,7 +163,10 @@ export default function SwapModalHeader({
               <Trans>To</Trans>
             </ThemedText.Body>
             <ThemedText.Body fontSize={14} color={theme.text3}>
-              <FiatValue fiatValue={fiatValueOutput} priceImpact={priceImpact} />
+              <FiatValue
+                fiatValue={fiatValueOutput}
+                priceImpactParams={{ priceImpact, error: undefined, loading: false }}
+              />
             </ThemedText.Body>
           </RowBetween>
           <RowBetween align="flex-end">
